Consolidate handler logic into subroutines
Add more handler functional tests.

Remove the negative content length logic since, on testing, the Go
stdlib never returns a negative content length. The original was getting
the negative length on a HEAD request. Since this code does not do the
pre-fetch HEAD, there is no need to be concerned with it.